+++ src/platform/packages/shared/kbn-discover-utils/moon.yml
@@ -41,8 +41,6 @@ dependsOn:
   - '@kbn/utility-types'
   - '@kbn/esql-language'
   - '@kbn/controls-schemas'
-  - '@kbn/share-plugin'
-  - '@kbn/saved-search-plugin'
   - '@kbn/control-group-renderer'
 tags:
   - shared-common
