+++ src/platform/packages/shared/kbn-discover-utils/tsconfig.json
@@ -37,8 +37,6 @@
     "@kbn/utility-types",
     "@kbn/esql-language",
     "@kbn/controls-schemas",
-    "@kbn/share-plugin",
-    "@kbn/saved-search-plugin",
     "@kbn/control-group-renderer"
   ]
 }
