+++ src/platform/plugins/shared/saved_search/common/types.ts
@@ -16,19 +16,10 @@ import type {
 } from '@kbn/data-plugin/common';
 import type { SavedObjectReference } from '@kbn/core-saved-objects-server';
 import type { SavedObjectsResolveResponse } from '@kbn/core/server';
-import type { SerializableRecord } from '@kbn/utility-types';
 import type { DataGridDensity } from '@kbn/unified-data-table';
 import type { SortOrder } from '@kbn/discover-utils';
 import type { DiscoverSessionTab as DiscoverSessionTabSchema } from '../server';
 import type { VIEW_MODE } from '.';
-
-export interface DiscoverGridSettings extends SerializableRecord {
-  columns?: Record<string, DiscoverGridSettingsColumn>;
-}
-
-export interface DiscoverGridSettingsColumn extends SerializableRecord {
-  width?: number;
-}
 
 export type VisContextUnmapped =
   | {
